Move createDefaultConfig into factory
